+++ packages/client-helpers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @parcnet-js/client-helpers
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @parcnet-js/client-rpc@1.1.1
+
 ## 1.0.1
 
 ### Patch Changes